bison: v2.1 → v3.0.0 helpers-native migration (plumbing-only)
Bison migrated from legacy runtime-1.0-era pattern (Python full-agency
scanner + openclaw cron + mcporter subprocess + Wolverine-pattern
internal create_position) to helpers-native runtime-1.1.0 pattern.
Plumbing-only — NO thesis change. v2.1 asset whitelist (BTC/ETH/SOL),
MIN_SCORE 11, 9-component scoring, direction waterfall, and
conviction-scaled margin tiers preserved verbatim.

Six-layer plumbing flip:
1. MCP transport: mcporter subprocess (2.5-5s cold-start) →
   SenpiClient.mcp_call() in-process HTTPS (~280ms).
2. Signal emit: scanner called create_position directly; producer
   now emits via push_signal() to runtime /signals. Runtime opens
   via LLM-gated bison_entry action with FEE_OPTIMIZED_LIMIT,
   passing marginUsd/leverage through from signal data for
   conviction-tier sizing.
3. Reentrancy: no v2 lockfile to replace (Bison v2.x had none);
   producer_daemon owns scanner_lock with stale-PID auto-recovery.
4. Scheduler: openclaw cron (5min) → producer_daemon(
   interval_seconds=300). Long-lived process.
5. Risk gates: Python MAX_DAILY_LOSS_PCT + dynamic-slots PnL-tiered
   cap + asset-cooldowns.json → declarative risk.guard_rails.
   state/trade-counter.json and state/asset-cooldowns.json now
   vestigial (can be deleted).
6. Exit fee: DSL exits MARKET (taker 0.045%) → FEE_OPTIMIZED_LIMIT
   (maker-first 0.015%, 60s ALO timeout, taker fallback). Entries
   keep ensure_execution_as_taker: false per v2.1 patience rule.

Also: v2.x shipped a hardcoded DSL state template in scanner output.
v3.0 makes the runtime YAML's dsl_preset the single source of truth
for DSL state — no scanner-side DSL template.

Dynamic daily-cap circuit breaker dropped: v2.1's tiered cap
(3 base / 6 hard cap by realized PnL) replaced by static
max_entries_per_day=3 + daily_loss_limit_pct=10 + drawdown_halt_pct=25.
Conservative ceiling and hard stop preserved; intermediate
PnL-aware tiers sacrificed for state-file-free reliability.

Files:
- bison/scripts/bison-producer.py (NEW; replaces bison-scanner.py)
- bison/scripts/bison_config.py (rewrite: drop mcporter subprocess,
  add SDK probe + lazy SenpiClient wrapper, mcp_call routes through
  SenpiClient, mcporter_call alias preserved)
- bison/runtime.yaml (add external_scanner bison_signals, LLM-gated
  bison_entry with conviction-tier marginUsd/leverage pass-through,
  risk.guard_rails, FEE_OPTIMIZED_LIMIT entries+exits; DSL preset
  preserved verbatim from v2.1 with time-cuts DISABLED)
- bison/SKILL.md (v3.0.0 frontmatter + architecture; thesis preserved)
- bison/README.md (helpers-native install — Step 0 openclaw.json,
  Step 1 install runtime skill, Step 2 pull files, Step 3 config.json,
  Step 4 env vars, Step 5 recreate runtime + launch daemon)
- bison/scripts/bison-scanner.py (deleted)
